% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fxn_httr}
\alias{fxn_httr}
\title{Use httr to do a GET request}
\usage{
fxn_httr(some_url = "http://httpbin.org/")
}
\arguments{
\item{some_url}{(character) a url}
}
\description{
Use httr to do a GET request
}
