% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_httr2.R
\name{fxn_httr2}
\alias{fxn_httr2}
\title{Use httr2 to do a GET request}
\usage{
fxn_httr2(some_url = "http://httpbin.org/")
}
\arguments{
\item{some_url}{(character) a url}
}
\description{
Use httr2 to do a GET request
}
